+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Update to Grafana 10.2.1 (#292)
 - Update ESLint configuration (#294)
 - Add Autosize Code Editor (#295)
+- Add Data Source option for Reset button (#296)
 
 ### Bugfixes